pyproject.toml
[tool.poetry]
name = "livraria"
version = "0.1.0"
description = "API Livraria com DRF"
authors = ["Rodrigo Cargnelutti <cargnelutti@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1"
gunicorn = "^20.1.0"
whitenoise = "^6.2.0"
djangorestframework = "^3.13.1"
flake8 = "^5.0.4"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
black = "^22.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
